+++ src/helpers/isModifierRequired.js
