+++ src/helpers/runModifiers.js
@@ -1,5 +1,5 @@
-import isFunction from './isFunction';
-import findIndex from './findIndex';
+import isFunction from '../utils/isFunction';
+import findIndex from '../utils/findIndex';
 
 /**
  * Loop trough the list of modifiers and run them in order, each of them will then edit the data object
